# docs/swagger.yaml
basePath: /v1/2024
definitions:
  model.AddNewContactMessageParams:
    properties:
      email:
        type: string
      message:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
...
  model.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  model.RegisterInput:
    properties:
      verify_key:
        type: string
      verify_purpose:
        type: string
      verify_type:
        type: integer
    type: object
  model.SetupTwoFactorAuthInput:
    properties:
      two_factor_auth_type:
        type: string
      two_factor_email:
        type: string
      user_id:
        type: integer
    type: object
  model.TwoFactorVerifycationInput:
    properties:
      two_factor_code:
        type: string
      user_id:
        type: integer
    type: object
  model.UpdateContactMessageParams:
    properties:
      status:
        type: integer
    type: object
...
  model.UpdatePasswordInput:
    properties:
      password:
        type: string
      token:
        type: string
    type: object
  model.VerifyInput:
    properties:
      verify_code:
        type: string
      verify_key:
        type: string
    type: object
  response.ErrResponse:
    properties:
      code:
        type: integer
      detail: {}
      messgae:
        type: string
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      messgae:
        type: string
    type: object
host: localhost:8082
info:
  contact:
    email: quang0706r@gmail.com
    name: TEAM QUANG
    url: https://github.com/Quangnguyen1505/Ecommerce-Go
  description: This is a server ecommerce.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/Quangnguyen1505/Ecommerce-Go
  title: Demo API Ecommerce
  version: 1.0.0
paths:
  /contact-messages:
    get:
      consumes:
      - application/json
      description: When admin wants to get all contact messages filtered by status
      parameters:
      - description: authorization token
        in: header
        name: authorization
        required: true
        type: string
      - description: x-client-id user
        in: header
        name: x-client-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Get all contact messages by status
      tags:
      - contactmessage
    post:
      consumes:
      - application/json
      description: When user sends a new contact message
      parameters:
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.AddNewContactMessageParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Create new contact message
      tags:
      - contactmessage
  /contact-messages/{id}:
    delete:
      consumes:
      - application/json
      description: When admin wants to delete a contact message
      parameters:
      - description: authorization token
        in: header
        name: authorization
        required: true
        type: string
      - description: x-client-id user
        in: header
        name: x-client-id
        required: true
        type: string
      - description: Contact Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Delete contact message
      tags:
      - contactmessage
    get:
      consumes:
      - application/json
      description: When admin wants to get a contact message by its ID
      parameters:
      - description: Contact Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Get contact message by ID
      tags:
      - contactmessage
    put:
      consumes:
      - application/json
      description: When admin wants to update status of a contact message
      parameters:
      - description: authorization token
        in: header
        name: authorization
        required: true
        type: string
      - description: x-client-id user
        in: header
        name: x-client-id
        required: true
        type: string
      - description: Contact Message ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UpdateContactMessageParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Update contact message
      tags:
      - contactmessage
...
  /users/login:
    post:
      consumes:
      - application/json
      description: When user is login save dbs
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: User Login
      tags:
      - account manager
  /users/register:
    post:
      consumes:
      - application/json
      description: When user is registered send otp to email
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: User Registrasion
      tags:
      - account manager
  /users/two-factor/setup:
    post:
      consumes:
      - application/json
      description: When user is login after 2 factor authentication
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.SetupTwoFactorAuthInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: User 2FA
      tags:
      - account 2fa
  /users/two-factor/verify:
    post:
      consumes:
      - application/json
      description: When user is login after enter code with pupose verify
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.TwoFactorVerifycationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: User Verify 2FA
      tags:
      - account 2fa
  /users/updatePass:
    post:
      consumes:
      - application/json
      description: When user is VerifyOTP ok after Update Password
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UpdatePasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: User Update Password
      tags:
      - account manager
  /users/verifyOTP:
    post:
      consumes:
      - application/json
      description: When user is register after verify otp
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.VerifyInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: User Verify OTP
      tags:
      - account manager
swagger: "2.0"